chore(platforms): rename cache mem-cache
