+++ platforms/src/index.ts
@@ -42,5 +42,5 @@ export const providers = new Providers([
 export * from "./types";
 export { Platform as PlatformClass } from "./utils/platform";
 export { platforms as platforms };
-export { initCacheSession, loadCacheSession, clearCacheSession } from "./utils/cache";
+export { initCacheSession, loadCacheSession, clearCacheSession } from "./utils/mem-cache";
 export { handleAxiosError } from "./utils/handleAxiosError";
