+++ platforms/src/utils/mem-cache.ts
